Moved Results to own file;
